refresh interfaces calamares
