--- src/interfaces/calamares/bootloader.schema.yaml
@@ -1,27 +1,0 @@
-# SPDX-FileCopyrightText: 2020 Adriaan de Groot <groot@kde.org>
-# SPDX-License-Identifier: GPL-3.0-or-later
----
-$schema: https://json-schema.org/schema#
-$id: https://calamares.io/schemas/bootloader
-additionalProperties: false
-type: object
-properties:
-    efiBootLoaderVar: { type: string }
-    efiBootLoader: { type: string }
-    kernelSearchPath:  { type: string }
-    kernelName:  { type: string }
-    kernelParams: { type: array, items: { type: string } }
-    kernelPattern: { type: string }
-    loaderEntries: { type: array, items: { type: string } }
-    refindKernelList: { type: array, items: { type: string } }
-
-    # Programs
-    grubInstall:  { type: string }
-    grubMkconfig:  { type: string }
-    grubCfg:  { type: string }
-    grubProbe: { type: string }
-    efiBootMgr: { type: string }
-
-    efiBootloaderId:  { type: string }
-    installEFIFallback: { type: boolean }
-    installHybridGRUB: { type: boolean }
