--- src/interfaces/calamares/cal34.bootloader.ts
@@ -1,25 +1,0 @@
-/* eslint-disable */
-/**
- * This file was automatically generated by json-schema-to-typescript.
- * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
- * and run json-schema-to-typescript to regenerate this file.
- */
-
-export interface HttpsCalamaresIoSchemasBootloader {
-  efiBootLoaderVar?: string;
-  efiBootLoader?: string;
-  kernelSearchPath?: string;
-  kernelName?: string;
-  kernelParams?: string[];
-  kernelPattern?: string;
-  loaderEntries?: string[];
-  refindKernelList?: string[];
-  grubInstall?: string;
-  grubMkconfig?: string;
-  grubCfg?: string;
-  grubProbe?: string;
-  efiBootMgr?: string;
-  efiBootloaderId?: string;
-  installEFIFallback?: boolean;
-  installHybridGRUB?: boolean;
-}
