--- src/interfaces/calamares/cal34.displaymanager.ts
@@ -1,34 +1,0 @@
-/* eslint-disable */
-/**
- * This file was automatically generated by json-schema-to-typescript.
- * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
- * and run json-schema-to-typescript to regenerate this file.
- */
-
-export interface HttpsCalamaresIoSchemasDisplaymanager {
-  /**
-   * @minItems 1
-   */
-  displaymanagers?: [
-    "slim" | "sddm" | "lightdm" | "gdm" | "mdm" | "lxdm" | "greetd",
-    ...("slim" | "sddm" | "lightdm" | "gdm" | "mdm" | "lxdm" | "greetd")[]
-  ];
-  defaultDesktopEnvironment?: {
-    executable: string;
-    desktopFile: string;
-    [k: string]: unknown;
-  };
-  basicSetup?: boolean;
-  sysconfigSetup?: boolean;
-  greetd?: {
-    greeter_user?: string;
-    greeter_group?: string;
-    greeter_css_location?: string;
-  };
-  lightdm?: {
-    preferred_greeters?: string[];
-  };
-  sddm?: {
-    configuration_file?: string;
-  };
-}
