--- src/interfaces/calamares/cal34.finished.ts
@@ -1,14 +1,0 @@
-/* eslint-disable */
-/**
- * This file was automatically generated by json-schema-to-typescript.
- * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
- * and run json-schema-to-typescript to regenerate this file.
- */
-
-export interface HttpsCalamaresIoSchemasFinished {
-  restartNowEnabled?: boolean;
-  restartNowChecked?: boolean;
-  restartNowCommand?: string;
-  restartNowMode?: "never" | "user-unchecked" | "user-checked" | "always";
-  notifyOnFinished?: boolean;
-}
