--- src/interfaces/calamares/cal34.packages.ts
@@ -1,39 +1,0 @@
-/* eslint-disable */
-/**
- * This file was automatically generated by json-schema-to-typescript.
- * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
- * and run json-schema-to-typescript to regenerate this file.
- */
-
-export interface HttpsCalamaresIoSchemasPackages {
-  backend:
-    | "apk"
-    | "apt"
-    | "dnf"
-    | "dnf5"
-    | "entropy"
-    | "luet"
-    | "packagekit"
-    | "pacman"
-    | "pamac"
-    | "portage"
-    | "yum"
-    | "zypp"
-    | "dummy";
-  update_db?: boolean;
-  update_system?: boolean;
-  skip_if_no_internet?: boolean;
-  pacman?: {
-    num_retries?: number;
-    disable_download_timeout?: boolean;
-    needed_only?: boolean;
-  };
-  operations?: {
-    install?: unknown[];
-    remove?: unknown[];
-    try_install?: unknown[];
-    try_remove?: unknown[];
-    localInstall?: unknown[];
-    source?: string;
-  }[];
-}
