--- src/interfaces/calamares/cal34.partition.ts
@@ -1,39 +1,0 @@
-/* eslint-disable */
-/**
- * This file was automatically generated by json-schema-to-typescript.
- * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
- * and run json-schema-to-typescript to regenerate this file.
- */
-
-export interface HttpsCalamaresIoSchemasPartition {
-  efiSystemPartition?: string;
-  efiSystemPartitionSize?: string;
-  efiSystemPartitionName?: string;
-  efi?: {
-    recommendedSize?: string;
-    minimumSize?: string;
-    label?: string;
-    mountPoint?: string;
-  };
-  lvm?: {
-    enable?: boolean;
-  };
-  userSwapChoices: ("none" | "reuse" | "small" | "suspend" | "file")[];
-  armInstall?: boolean;
-  allowZfsEncryption?: boolean;
-  drawNestedPartitions?: boolean;
-  alwaysShowPartitionLabels?: boolean;
-  defaultFileSystemType?: string;
-  availableFileSystemTypes?: string[];
-  mountpointFilesystemRestrictions?: unknown[];
-  luksGeneration?: "luks1" | "luks2";
-  enableLuksAutomatedPartitioning?: boolean;
-  preCheckEncryption?: boolean;
-  essentialMounts?: string[];
-  allowManualPartitioning?: boolean;
-  showNotEncryptedBootMessage?: boolean;
-  partitionLayout?: unknown[];
-  initialPartitioningChoice?: "none" | "erase" | "replace" | "alongside" | "manual";
-  initialSwapChoice?: "none" | "small" | "suspend" | "reuse" | "file";
-  requiredStorage?: number;
-}
