--- src/interfaces/calamares/finished.schema.yaml
@@ -1,13 +1,0 @@
-# SPDX-FileCopyrightText: 2020 Adriaan de Groot <groot@kde.org>
-# SPDX-License-Identifier: GPL-3.0-or-later
----
-$schema: https://json-schema.org/schema#
-$id: https://calamares.io/schemas/finished
-additionalProperties: false
-type: object
-properties:
-    restartNowEnabled: { type: boolean, default: true }  # TODO:3.3: remove
-    restartNowChecked: { type: boolean, default: false } # TODO:3.3: remove
-    restartNowCommand: { type: string }
-    restartNowMode: { type: string, enum: [ never, user-unchecked, user-checked, always ] }
-    notifyOnFinished: { type: boolean }
