+++ src/interfaces/calamares/i-calamares-bootloader.ts
@@ -5,7 +5,7 @@
  * and run json-schema-to-typescript to regenerate this file.
  */
 
-export interface HttpsCalamaresIoSchemasBootloader {
+export interface ICalamaresBootloader {
   efiBootLoaderVar?: string;
   efiBootLoader?: string;
   kernelSearchPath?: string;
