+++ src/interfaces/calamares/i-calamares-displaymanager.ts
@@ -6,13 +6,29 @@
  * license: MIT
  */
 export interface ICalamaresDisplaymanager {
-    displaymanagers: string [];
-    basicSetup: boolean;
-    sysconfigSetup: boolean;
-    defaultDesktopEnvironment?: DesktopEnvironment;
-}
-
-export interface DesktopEnvironment {
+  /**
+   * @minItems 1
+   */
+  displaymanagers?: [
+    "slim" | "sddm" | "lightdm" | "gdm" | "mdm" | "lxdm" | "greetd",
+    ...("slim" | "sddm" | "lightdm" | "gdm" | "mdm" | "lxdm" | "greetd")[]
+  ];
+  defaultDesktopEnvironment?: {
     executable: string;
     desktopFile: string;
+    [k: string]: unknown;
+  };
+  basicSetup?: boolean;
+  sysconfigSetup?: boolean;
+  greetd?: {
+    greeter_user?: string;
+    greeter_group?: string;
+    greeter_css_location?: string;
+  };
+  lightdm?: {
+    preferred_greeters?: string[];
+  };
+  sddm?: {
+    configuration_file?: string;
+  };
 }
