+++ src/interfaces/calamares/i-calamares-packages.ts
@@ -6,25 +6,34 @@
  * license: MIT
  */
 export interface ICalamaresPackages {
-    backend: 'apk' | 'apt' | 'dnf' | 'dnf5' | 'entropy' | 'luet' | 'packagekit' | 'pacman' | 'pamac' | 'portage' | 'yum' | 'zypp' | 'dummy';
-    update_db?: boolean;
-    update_system?: boolean;
-    skip_if_no_internet?: boolean;
-    pacman?: PacmanConfig;
-    operations?: Operation[];
-}
-
-export interface PacmanConfig {
+  backend:
+    | "apk"
+    | "apt"
+    | "dnf"
+    | "dnf5"
+    | "entropy"
+    | "luet"
+    | "packagekit"
+    | "pacman"
+    | "pamac"
+    | "portage"
+    | "yum"
+    | "zypp"
+    | "dummy";
+  update_db?: boolean;
+  update_system?: boolean;
+  skip_if_no_internet?: boolean;
+  pacman?: {
     num_retries?: number;
     disable_download_timeout?: boolean;
     needed_only?: boolean;
-}
-
-export interface Operation {
-    install?: any[];
-    remove?: any[];
-    try_install?: any[];
-    try_remove?: any[];
-    localInstall?: any[];
+  };
+  operations?: {
+    install?: unknown[];
+    remove?: unknown[];
+    try_install?: unknown[];
+    try_remove?: unknown[];
+    localInstall?: unknown[];
     source?: string;
+  }[];
 }
